Finished GL_TEXTURE_2D_ARRAY support

+++ src/gl_glsl.h
@@ -13,7 +13,7 @@ int clampi(int i, int min, int max);
 vec4 texture1D(GLuint tex, float x);
 vec4 texture2D(GLuint tex, float x, float y);
 vec4 texture3D(GLuint tex, float x, float y, float z);
-vec4 texture2DArray(GLuint tex, float x, float y, float z);
+vec4 texture2DArray(GLuint tex, float x, float y, int z);
 vec4 texture_cubemap(GLuint texture, float x, float y, float z);
